copperplate/registry: Make inter-user sharing configurable
There are two scenarios to consider when it comes to sharing registry
mounts: either Xenomai applications only run as root or a group was
given the rights to run such applications as well. In the former, case
there is no need to expose registry data to non-root users. The latter
case still requires more work and is currently not supported /wrt the
registry. But that will only require sharing of the anonymous session.

To prepare controlled sharing of registry mounts, make them
configurable. This is done by a new command line option --shared for
sysregd. By default, sharing is off in order to avoid unwanted and
needless information disclosure about application internals to
unprivileged users.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>