scripts: Add multi-user service setup scripts
This script can be hooked up with the system in order to allow
unprivileged members of the xenomai group to access real-time services
as well. It configures the module parameter, creates the registry root
directory with the appropriate access rights and ensures that a sysreg
daemon for the anon session is running.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>