hal/arm: simplify fpu handling
Currently, when returning to root domain while fpu context has not been
altered by real-time domain, fpu is unconditionnaly re-enabled on UP
systems. This causes a problem starting with kernel 3.16 which handles
FPU faults with irqs on: if the fpu is reenabled on this path, Linux
assumes that the fault can not be caused by the fpu not being enabled
resulting in a fallback to the "undefined instruction" handler.

To avoid this, test whether fpu is enabled when leaving the root domain
and restore the fpu context exactly as it was before the preemption by
the real-time domain.