copperplate/threadobj: mercury: work around glibc's PI bug with condvars
We have a problem with all glibc releases to date with condition
variables, which may suffer from priority inversion:
https://sourceware.org/bugzilla/show_bug.cgi?id=11588
https://lwn.net/images/conf/rtlws11/papers/proc/p10.pdf

--enable-condvar-workaround enables code to boost threads temporarily
prior to signaling or waiting for an event on a condition variable
protected by a PI mutex. Although this may be fairly costly, this
prevents nasty CPU starvation issues which may be otherwise caused by
priority inversions while blocking on the condvar's internal mutex
as implemented by the glibc.

This bug only affects Mercury-based applications which depend on
strict PI to prevent from harmful priority inversion issues. The
Cobalt implementation of condvars is PI-aware and needs no work
around.