nucleus/registry: initialize vfile structure to 0
The vfile API expects the structures it receives to be fully
initialized. While structures with static storage are automatically
initialized to 0, the vfile structure passed by the registry to the
vfile API is part of a dynamically allocated structure. So, initialize
this structure to 0.

This fixes the following bug:
[23726.087793] Xenomai: bug at kernel/xenomai/nucleus/vfile.c:320
(it->vfile->entry.refcnt < 0)

Obtained when reading native semaphores exported proc entry with Xenomai
compiled with debugs on.