cobalt/posix: Add support for return long values from syscalls
So far, handle_head/root_syscall limited to width of return values to
int, thus 32-bit. This was waiting to break, and it did over the rework
of the mayday syscall which now passed a pointer back. But also I/O
syscalls were vulnerable on 64-bit platforms, at least in principle.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>