cobalt/sched: do not choke on redundant policy setting
Setting the base scheduling class of a thread to the current value
should not cause such thread to be declared again to the policy
module. It is not only useless, but declaration handlers are not
expected to cope with such situation.

See http://www.xenomai.org/pipermail/xenomai/2015-March/033563.html.