cobalt/posix: fix potential race when reclaiming global resources
In theory, a Xenomai process on a different CPU might race with the
reclaim code if it starts then exits while the global resources are
being released on the current CPU. Unlikely, but possible though.

Queue the global resources to be reclaimed from the exiting process to
a local stash under nklock protection before cleaning them up.