lib: move early setup code to the boilerplate layer
This is a preparation step to unify the Xenomai setup interface for
both pure Cobalt/POSIX and Copperplate-based applications.

The goal is to expose a common set of basic command line switches
which apply to all application programs regardless of the API being
used (e.g. --dump-config), and a single setup call for bootstrapping
the core libraries in sequence.