boilerplate, copperplate: introduce API to manipulate tunables
The core libraries define a set of tunable parameters, this commit
provides an API for applications to assign and fetch their current
value.

Parameters are divided in 'config' and 'runtime' tunable sets. The
former may be assigned new values until the core starts running the
->init() handlers from the setup descriptors, at which point any
attempt to change them triggers an assertion when --enable-assert is
in effect. The latter can be changed at any point in time.

A single API is provided in order to frame all possible manipulations
within a small set of generic helpers, so that we can do basic error
detection at build time and runtime, and keep options open for later
extending this API, such as automated tracing and reporting.