copperplate/heapobj-malloc: fix locking
Do not call malloc/free while potentially holding a RT mutex: Cobalt
might send us a SIGDEBUG if we happen to switch to secondary mode
there.

In addition, malloc(3) is not a cancellation point, so there is no
need for cleanup handlers. The caller should have entered a deferred
cancellation section prior to calling us.