lib/cobalt: drop "no shadow" binding mode
This mode was meant to allow application programs to dlopen() the
legacy Xenomai core library (aka libxenomai.so), in which the services
used to be bootstrapped from an embedded library constructor, binding
to the real-time core, before the current thread got shadowed
automatically. The no-shadow mechanism was meant at preventing the
latter action from taking place at the end of the init sequence.

In the current implementation, the services are bootstrapped by an
explicit call to xenomai_init(), usually done from a wrapper
interposing on the main() routine, not from a library constructor
anymore. Therefore, dlopen()'ing libcobalt.so per se will not cause
the calling context to be shadowed. Calling xenomai_init() will be
required to activate the Xenomai services for the calling process at
some point, however the program is free to choose the context from
which this should be done.

In parallel, running over a shadowed context is required for
traversing the init code of the various libraries bootstapped on
behalf of xenomai_init().

So, basically we don't support "no shadow" binding anymore, but this
is a non-issue since it became useless.