boilerplate: declare missing libc functions when wrapping
When compiling with a toolchain which does not implement
pthread_setname_np or clock_nanosleep, these functions are not declared
as a result of including their respective header.

However, when passing --skin=posix to xeno-config, an application
expects these functions to be declared, since they are wrapped and
implemented by libcobalt.

Fix this by moving their declaration in boilerplate/libc.h out of the
ifdef __IN_XENO__ section, but making it depend on a __COBALT_WRAP__
symbol, set in xeno-config when --skin=posix is passed.