cobalt/sched: avoid leaking the reschedule bit
handle_setaffinity_event ends up calling migrate_thread() which sets
the rescheduling bit. However, since this happens for a passive
migration, xnsched_run() is not called, and the reschedule bit
remains. This causes remote notifications to not send an IPI, and
delay them until the system calls xnsched_run() for another reason.

Fix this by setting the reschedule bit in xnsched_migrate_passive() only
if the thread blockbits are not set.