cobalt/timer: avoid early tick in case of overrun
There is a race condition in the current timer code when a timer period
is close to the maximum jitter. It may happen that the next timer
interrupt is already pending while xntimer_get_overruns computes the
count of overruns. In that case, the count of overruns returned to
user-space is one too much, and the application ends up expecting the
next tick one period later than when it will actually happens.

Avoid that problem by advancing the timer in xntimer_get_overruns.