cobalt/clock: introduce CPU affinity for clock event sources
Xenomai clocks may be backed by devices issuing event notifications on
a restricted set of CPUs only. Typically, the interrupt distributor of
GIC-equiped ARM processors is programmed to route shared processor
interrupts to CPU0 by default. For instance, a Xenomai clock paced by
a GPIO input signal would tick exclusively on CPU0 in that
configuration.

We have to make sure that timers are always attached to per-CPU queues
which are expected to receive event notifications from the underlying
clock device, otherwise threads may wait for events indefinitely on
those timers.

Previously, a restrictive assumption was made that all CPUs from the
real-time set should be able to receive events, indirectly assuming
that each ticking device is a per-CPU device. Although this is correct
for the core Xenomai clock based on local timers for SMP, this turns
out to be wrong quite frequently with external Xenomai clocks backed
by common devices issuing a periodic master beat.

To fix this, we require the set of CPUs the event source may tick on
to be declared as part of the Xenomai clock properties. This
information is then used to migrate timers to an appropriate CPU if
need be, and to detect mismatching CPU affinities between timers and
clock devices when DEBUG_COBALT is enabled.