cobalt/thread: force secondary mode for joining threads
Make xnthread_join() switch the caller to secondary mode prior to
waiting for the target thread termination. The original runtime mode
is restored upon return.

Since the joiner was already synchronized on an event that may be sent
by the joinee from secondary mode exclusively, this change does not
drop any real-time guarantee for the joiner: there has never been any
in the first place.

This is a preparation step to a stricter synchronization between the
joiner and the joinee, especially in the SMP case.