lib/cobalt: simplify checks for secondary mode
Whether a current thread is currently running in primary/secondary
mode can be checked by inspecting the status of the XNRELAX bit in the
thread's u_window. The lack of u_window denotes a non-Xenomai, regular
thread which for which the current mode is always relaxed.

A thread's handle and its u_window area are tightly coupled: the
latter has to be valid for any thread with a valid handle.

Introduce cobalt_is_relaxed() for testing the current mode, and fixup
all open coded versions of this routine.

In addition, assert_nrt_fast() is marked as deprecated, since
assert_nrt() is not a slow call anymore.