cobalt/rtdm: allow setting the kernel device class of a driver
It may be convenient to set the Linux device class for a particular
RTDM driver, either to provide a specific set of sysfs attributes for
the attached devices, or using a different hierarchy for the device
nodes (instead of /dev/rtdm).

The new rtdm_drv_set_sysclass() call allows this, provided the
non-default class is set before the first device referring to the
driver is registered by a call rtdm_dev_register().