testsuite/switchtest: detect more FPU errors.
Do not store the same value in all FP registers, so that if we access
the backup area with an offset, we can detect it.

Recheck the fpu registers after mode switches, in order to verify that
the Linux and Xenomai backup areas are identical for Xenomai threads.
This would have allowed to detect the issue fixed by commit
d4e755b2a9909afb7bbd0a522ff1d97718494cd7