debian: remove device nodes from libxenomai1
The library package does not seem to be the correct place for device
nodes. People might install the lib for cross compilation and they will
not need the nodes. If a non-root user extracts the package
node-creation will fail.

On modern systems using debian packages the device nodes will be
generated by udev. And on legacy systems it should probably have been a
bunch of mknods in a post-install step instead of packaging the nodes.

Signed-off-by: Henning Schild <henning.schild@siemens.com>