boilerplate: allow for auto-init of external shared libraries
3rd-party shared libraries for which Xenomai auto-initialization is
required must include a position-independent version of the bootstrap
module.

This means that we need to export a PIC object such libraries can
include, in addition to a non-PIC object to be linked in regular
executables as usual. Such objects are available as
@libdir@/xenomai/bootstrap-pic.o and @libdir@/xenomai/bootstrap.o
respectively.

xeno-config* scripts now accept the --auto-init-solib option to emit
the proper link directives for including the PIC bootstrap module into
shared libraries when --ldflags is given.