boilerplate: drop __real_main() wrapper
Such kind of wrappers is supposed to help when --wrap is not given at
link time, in resolving direct references to the real symbols such as
those which may be done from the __wrap routines.

This said, __real_main() is special, in that there is no reason to
call such routine explicitly if --wrap is not given, by
definition. Only the bootstrap module (init/bootstrap.c) may call the
real main() entry point this way, and such module is glued to the
executable only if --wrap is passed to the linker.

Therefore we may drop this particular wrapper and the file which only
contains it altogether.