cobalt/timer: consider clock event affinity when pinning timers
Pinning a timer to the same CPU than the caller is affine to must take
into account the affinity of the clock device backing the timer.

When an affinity mismatch is detected by xntimer_set_sched(), the
timer affinity is left unchanged, which will certainly cause
additional overhead due to the requirement of waking up a thread on a
remote CPU from the timer's handler, but will also keep the situation
sane.