include/psos: remove dependency on tunables.h
The psos.h file pollutes the application's namespace with defines
of the boilerplate code because the tunables.h is included. In our
application we already have a definition of the list_entry macro,
resulting in a redefinition. The psos.h should be kept clean.

Following small test program clearly shows the issue:

int main(int argc, char * const argv[])
{
  return 0;
}

When compiling the code above a warning is generated:

simple_test.c:3:1: warning: "list_entry" redefined
In file included from boilerplate/list.h:24,
                 from boilerplate/setup.h:21,
                 from boilerplate/tunables.h:22,
                 from psos/psos.h:28,
                 from simple_test.c:1:
boilerplate/shared-list.h:233:1: warning: this is the location of the previous definition

This patch resolves the issue by moving the tunable related functionality in
a seperate header file.