lib/boilerplate: bootstrap: fix Blackfin build
The bootstrap module object is built in two forms:

- in static object form, to be glued to the main executable, which
  should include a wrapper interposing on the main() routine for
  auto-init purpose. Such wrapper is activated when symbol wrapping is
  enabled at link time (--wrap).

- in dynamic object form, to be included in a shared library target
  which enables the auto-init feature. This form should not include
  any wrapper to a main() routine - which does not exist - but only a
  constructor routine performing the inits.

The FDPIC format is used for building static and dynamic objects on
the Blackfin architecture, for which GCC defines __PIC__ in both
cases. This prevents from testing __PIC__ as a way to differentiate
between the static and dynamic versions of the bootstrap module on
Blackfin.

To fix this, set and check for a dedicated macro (__INTERCEPT_MAIN__)
to figure out whether the main() interception code should be present
in the relocatable object.