boilerplate/tlsf: work around excessive meta-data overhead for main pool
e0e6ae76 caused a regression with failure to initialize the main pool
with TLSF, uncovering an issue in the allocator code which fails to
account properly for the overhead induced by the meta-data.

Overhead footprints and management is a recurring problem with TLSF,
particularly on 64bit platforms; for this reason there is a plan for
dropping TLSF entirely from the Xenomai code base. Until this happens,
the current change works around the issue to keep things going.