boilerplate/setup: allow DSO bootstrap modules to coexist with main one
Gluing a PIC bootstrap module to a DSO is a common way to start the
Xenomai services before constructors of static C++ objects
instantiated by the DSO and which depend on such services, are
invoked.

With such module present in the DSO, a Cobalt-based application
process would attach itself to the Cobalt core via the bootstrapping
sequence, enabling the real-time services for the whole process,
including the static C++ constructors which may be invoked when
attaching the shared object.

e.g. in libfoo.so:

class FOO {
protected:
	sem_t sem;
public:
	static FOO bar;
};

...

FOO FOO::bar;

FOO::FOO()
{
	// Initialize a Cobalt semaphore
	sem_init(&sem, 0, 0);
}

However, adding a bootstrap module to a DSO currently prevents from
merging a second bootstrap module aimed at kickstarting the main
program's setup code and early inits, due to conflicting symbols and
the restriction on invoking xenomai_init() only once for any given
process.

The changes introduce a DSO-specific initialization service
(xenomai_init_dso()), which may be called from the context of the
dynamic linker attaching a shared object to the current
executable. The DSO-specific bootstrap module - if present - will call
xenomai_init_dso(), there is no restriction on the number of DSOs
calling this service.

The non-DSO bootstrap module invokes xenomai_init() as previously, for
kickstarting the main program's early inits.