cobalt/x86: provide generic version of strncpy_from_user_nocheck()
The former open-coded version of this routine caused havoc in the
exception table generated for recent kernels after the introduction of
a new format for exception entries during the 4.5->4.6 time frame.

Switch this code to a generic form which does not need to know about
the implementation details of exception fixup code.