scripts/xeno-config: add -Wl,--no-as-needed
Linking with "--as-needed" could cause the linker to not list essential
wrapping libraries as needed for a shared library or a dynamically
linked binary. The result would be runtime ld.so errors for the binaries
or linking problems with the .so.
gcc defaults to "--no-as-needed" that is why we never needed to specify
that flag explicitly. But gcc from Ubuntu is patched to default to
"--as-needed".
This patch enforces "--no-as-needed", which will have no effect on most
toolchains. On Ubuntu it will override the default and make sure xenomai
applications get linked as expected.

Signed-off-by: Henning Schild <henning.schild@siemens.com>