lib/copperplate: drop useless wrapper to main()
Referencing __real_main() only makes sense from the bootstrap module
(boilerplate/init), which is glued to a DSO/final executable when
symbol wrapping is in effect, exclusively.

The wrapper found in copperplate is a left over from a former
implementation of the auto init feature, which may cause an undefined
reference to main() when loading libcopperplate.so dynamically. Let's
drop it.