cobalt: Do not destroy info flags on remote thread suspension
Scenario: A high prio thread is running and a low-prio is waiting for a
timeout. Now the timeout occurs, and the low-prio thread is woken up but
can't run yet. If the high-prio thread suspends the low-prio one,
XNTIMEO will be lost due to that. Similar scenarios a possible with
other info flags and remote suspension reasons.

Address them by clearing info flags only if the current thread is
suspending itself.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>