copperplate/heapobj: tlsf: Fix unneeded pre-commit of memory pool when CONFIG_XENO_PSHARED set
When CONFIG_XENO_PSHARED is set, heapobj_pkg_init_private() was causing unneeded pages to be locked into memory by performing a temporary allocation of the size of the entire memory pool.

This patch reduces the size of this temporary allocation to a single page when CONFIG_XENO_PSHARED is set.

Signed-off-by: Charles Kiorpes <ckiorpes@gmail.com>