cobalt/rtdm: Split rtdm_fd_enter up
split rtdm_fd_enter, move the functionality where we store the fd
until after the open() call succeeds.  Calls where open() fail a fd is
left in the tree even after the cleanup code is executed.  If this fd
number is used again we will fail the call to open until a different
fd is used.  This patch addresses this situation by not adding the fd
into the tree until open has succeeded and the fd is valid.