drivers/gpio: Request gpio at open()
For the zynq platform (and possibly others in the future) we need to modify
gpio-core to request gpio pins before using them. The open() function will
now request a gpio and fail if it's already reserved. This should make
the pin request transparent to the user. The ability to request and release
pins is also available in an ioctrl message.  Tested on Microzed Zynq-7010
platform and the raspberry pi 2 board.