boilerplate/bootstrap: dual-license the bootstrap source code
lib/boilerplate/bootstrap.c is meant to be glued to the application
executable for bootstrapping the Cobalt services automatically at load
up.

From the standpoint of its sole author (i.e. me), doing so adds no
requirement with respect to the licensing terms of the containing
application.

Allow lib/boilerplate/bootstrap.c to be distributed under the terms of
the MIT license to reflect this fact.