drivers/gpio: allow open coded IRQ handling
Some client code may want to deal with GPIO interrupt handling
directly, without support for an irq_chip controller paired with the
GPIO chip.

Assume that a failure to retrieve the IRQ number associated to a pin
means exactly that, offloading interrupt handling to an open coded
implementation the client code should provide for.

In parallel, introduce rtdm_gpiochip_post_event() for notifying the
GPIO core from an incoming interrupt event on a given pin.