drivers/gpio: pass back allocated GPIO chip descriptor
rtdm_gpiochip_alloc() is a way to instantiate a single GPIO descriptor
instead of mapping all GPIO controllers defined by the firmware which
belong to a particular type. This routine is matched by
rtdm_gpiochip_remove() for deallocation.

We need the former to return a pointer to the allocated descriptor we
may then pass to the latter.