net/stack: restore original deletion sequence for service tasks
Restore the original sequence for deleting an internal service task
that should most likely be waiting for an input event: first destroy
the event resourcen, which should trigger immediate return from the
wait call with -EIDRM, then send a cancellation request via
rtdm_task_destroy() for good measure to exit the work loop if the task
was not aslept on that event.

This obviously assumes that all callers waiting for events do check
for the return value, as they must do.