boilerplate/setup: introduce destructors for __setup_call
Whit dlclose() we can actually invalidate the setup_list, it might end
up containing pointers to memory that got unmapped with dlclose(). The
next dlopen() will then cause a segfault.
Add destructors to that such referrences are remove from the setup_list
when calling dlclose().
The previous commit added a testcase that fails because of that issue,
with this commit the test runs fine.

Signed-off-by: Henning Schild <henning.schild@siemens.com>