smokey/posix-fork: actually wait for the process to exit
The comment says that it would be enough to just trigger the chain of
events. This test did not even fail if the binary to execute was not
there, not sure how much of the chain was exercised.
This patch makes the parent wait for what actually happens and should
return errors, should they happen. It prepares for another patch using
fork/exec.

Signed-off-by: Henning Schild <henning.schild@siemens.com>