cobalt: unconditionally allow idling requests
Wait for the idling interface rework from I-pipe/4.14, which will
provide more information for determining whether Cobalt should be ok
with entering the target idle state.

As a result of this change, the original kernel behavior is restored
for all ipipe-4.9.y patches with respect to entering an idle state,
including for the releases lacking commits #89146106e8 or #8d3fa22c95.

This change only affects kernels built with CONFIG_CPU_IDLE enabled.

NOTE: XNIDLE is intentionally kept for future use in the Cobalt core.