cobalt/posix: Harden cobalt_thread_shadow against invalid requests
If the caller has already a Xenomai shadow, we will crash rather than
failing gracefully. The reason is that pthread_discard expects secondary
mode - which may not be the case if the caller is already able to
migrate. So check early and avoid that we can run into this trap.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>