lib/cobalt: Avoid requesting shadowing for remote threads
Fix pthread_setschedparam_ex that it does not suggests the core to map
the current thread if the target thread of the request is actually not
pthread_self(). The core will reject such requests now, but we should
avoid them in the first place so that we get a clear error code.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>