cobalt: Fix error code of cobalt_thread_setschedparam_ex
This aligns the return code with cobalt_sched_setschedparam_ex: If there
is no u_winoff and no thread, we should return ESRCH so that userspace
can easily detect that it should call the standard libc service instead.
Fix the core and remove the related workaround from the userspace lib.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>