lib/cobalt: Pass setsched requests for unknown threads to glibc
This aligns the implementation with the documentation: If we get an
error from the kernel that suggests the thread is unknown to it, let
glibc deal with it.

Note that pthread_setschedparam_ex currently returns an invalid error
code, EINVAL rather than ESRCH, so that we try to detect the remote
thread case with some heuristic.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>